Тесты

+++ src/Controller/CourseController.php
@@ -28,10 +28,12 @@ class CourseController extends AbstractController
 
     /**
      * @Route("/new", name="app_course_new", methods={"GET", "POST"})
-     * @IsGranted("ROLE_SUPER_ADMIN", statusCode=403 , message="Нет доступа!")
      */
     public function new(Request $request, CourseRepository $courseRepository): Response
     {
+        if (!in_array('ROLE_SUPER_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_course_index');
+        }
         $course = new Course();
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
@@ -59,10 +61,12 @@ class CourseController extends AbstractController
 
     /**
      * @Route("/{id}/edit", name="app_course_edit", methods={"GET", "POST"})
-     * @IsGranted("ROLE_SUPER_ADMIN", statusCode=403 , message="Нет доступа!")
      */
     public function edit(Request $request, Course $course, CourseRepository $courseRepository): Response
     {
+        if (!in_array('ROLE_SUPER_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_course_index');
+        }
         $form = $this->createForm(CourseType::class, $course);
         $form->handleRequest($request);
 
@@ -79,11 +83,14 @@ class CourseController extends AbstractController
 
     /**
      * @Route("/{id}", name="app_course_delete", methods={"POST"})
-     * @IsGranted("ROLE_SUPER_ADMIN", statusCode=403 , message="Нет доступа!")
      */
     public function delete(Request $request, Course $course, CourseRepository $courseRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
+        if (!in_array('ROLE_SUPER_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_course_index');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $course->getId(), $request->request->get('_token'))) {
             $courseRepository->remove($course);
         }
